Speed up the PageIterator by evaluating items once per page (#572)
The `firebase_admin.auth.list_users().iterate_all()` method uses a sub-class
of PageIterator, which happens to access the .items computed property more
than once for every page of results. This has been changed so we take care
not to access the `self.items` property more than once per page.

This is a lot faster.