Added new app option clockSkewInSeconds
This option value is used for the token verification instead of the fixed 60 seconds from
the earlier commit.

This way, the user of firebase_admin can decide if he/she wants to set that value or not.
Also all existing uses of firebase_admin won't suddenly change behaviour, since if the
option is not specified, it's default of 0 is equivalent to what was used before the
introduction of the new option.